test addMember

+++ nodejs/app.js
@@ -177,10 +177,10 @@ app.post('/addMember', function (req, res) {
     console.log("test");
     const email = req.body.email;
     const verificationCode = req.body.verificationCode;
-    console.log("Rnunning /addMember "+email+" "+verificationCode);
+    console.log("Rnunning /addMember " + email + " " + verificationCode);
 
     //validate email before doing anything else
-    if (email!=null && email.endsWith("@peddie.org") && validator.validate(email)) {
+    if (email != null && email.endsWith("@peddie.org") && validator.validate(email)) {
         const username = email.substring(0, email.lastIndexOf("@"));
         var con = mysql.createConnection({
             host: "localhost",
@@ -221,13 +221,17 @@ app.post('/addMember', function (req, res) {
         const sourcePath = path.join(__dirname, '..', 'members', 'user-images', 'temp', username);
         const destPath = path.join(__dirname, '..', 'members', 'user-images', username);
 
-        fs.rename(sourcePath, destPath, (err) => {
-            if (err) {
-                console.log(`Error moving file: ${err}`);
-            } else {
-                console.log(`File moved successfully from ${sourcePath} to ${destPath}`);
-            }
-        });
+        if (fs.existsSync(sourcePath)) {
+            fs.rename(sourcePath, destPath, (err) => {
+                if (err) {
+                    console.log(`Error moving file: ${err}`);
+                } else {
+                    console.log(`File moved successfully from ${sourcePath} to ${destPath}`);
+                }
+            });
+        } else {
+            console.log(`Error: Source file does not exist: ${sourcePath}`);
+        }
 
     } else {
         res.send({ "error": true, "message": 'Invalid Email' });
